+++ src/main/java/fr/cesi/contractor/Repository/QuoteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuoteRepository extends JpaRepository<Quote, Long> {
+public interface QuoteRepository extends JpaRepository<Quote, Integer> {
 }
